Overhauled the contact map selection and part listening class.  Editors opening and closing are handled in much smoother way.  All of the known bugs in the contact map except for the tooltips issues have been fixed (update issues, redraw issues, issues with closing the last cmap, memory leaks, etc).
Also added new problem view feature.  Instead of using  a console for the modeling phase, the user now has a table that shows all of the errors that are in currently open files.  The file labels are still too long, but I will fix that later. 

Also completed the functionality for creating new projects and updated the images for the file explorer context menu.  

Cleaned some old files that were no longer used.

Updated errout filters for ANTLR so that the errors are parsed and converted to BNGLModel structures that are open for the whole workspace to use through property change listeners.

Updated the way that the abstract syntax tree is updated so that less code is used and it's more efficient. 

Removed old debugging statements.